requirements.txt
aniso8601==6.0.0
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
clickclick==1.2.2
cryptography==2.6.1
Flask==1.0.2
Flask-Jsonpify==1.5.0
Flask-SQLAlchemy==2.3.2
functools32==3.2.3.post2
idna==2.8
inflection==0.3.1
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
MarkupSafe==1.1.1
openapi-spec-validator==0.2.6
pathlib==1.0.1
pyasn1==0.4.5
pycparser==2.19
PyJWT==1.4.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.3
typing==3.6.6
urllib3==1.24.2
Werkzeug==0.15.2
